refactor : integer 타입 자료형으로 전체 수정

+++ src/main/java/lotto/domain/lottoTicket/LottoTicket.java
@@ -51,7 +51,7 @@ public class LottoTicket {
 	}
 
 	private void validateDuplication(List<LottoNumber> lottoNumbers) {
-		long distinctSize = new HashSet<>(lottoNumbers).size();
+		int distinctSize = new HashSet<>(lottoNumbers).size();
 
 		if (distinctSize != lottoNumbers.size()) {
 			throw new InvalidLottoTicketException(InvalidLottoTicketException.DUPLICATION);
